[~] decomposition
